spmedge/requirements.txt
# Core Dependencies
openai>=0.27.0
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
tqdm>=4.64.0

# Database & Data Handling
SQLAlchemy>=1.4.40
pandas>=1.5.0
numpy>=1.23.0

# Web Framework & API
flask>=2.2.0
flask-cors>=3.0.10

# Utilities
python-dotenv>=0.20.0
pydantic>=1.10.0
loguru>=0.6.0

# Testing
pytest>=7.0.0

# Data Parsing & Visualization
matplotlib>=3.5.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
html5lib>=1.1